reorg da reader files, add da sync mod
